+++ aioscraper/_helpers/http.py
@@ -13,7 +13,7 @@ from aioscraper.types import QueryParams, RequestCookies
 def parse_url(url: str, params: QueryParams | None) -> URL:
     parsed_url = URL(url)
     if params:
-        return parsed_url.update_query(params)
+        return parsed_url.extend_query(params)
 
     return parsed_url